UMDCTF2021/Web/return-of-the-flag-bay/web1/code/requirements.txt
Click
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug
Flask-Session
Flask-Cors
mysql-connector-python